requirements.txt
Django>=5.0
djangorestframework
openpyxl
